requirements.txt
google-generativeai
langchain_community
pypdf
chromadb
unstructured
langchain-google-genai
json
jq